Break ServiceAlarmBuilder into two
So that in future parts can be swapped out without reimplementing loads of boilerplate
