--- Watchman.Engine/Generation/ServiceAlarmBuilder.cs
@@ -1,188 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Watchman.AwsResources;
-using Watchman.Configuration;
-using Watchman.Configuration.Generic;
-
-namespace Watchman.Engine.Generation
-{
-    public class ServiceAlarmBuilder<T, TAlarmConfig>
-        where T:class
-        where TAlarmConfig : class, IServiceAlarmConfig<TAlarmConfig>, new()
-    {
-        private readonly IResourceSource<T> _tableSource;
-        private readonly IAlarmDimensionProvider<T> _dimensions;
-        private readonly IResourceAttributesProvider<T, TAlarmConfig> _attributes;
-
-        public ServiceAlarmBuilder(
-            IResourceSource<T> tableSource,
-            IAlarmDimensionProvider<T> dimensionProvider,
-            IResourceAttributesProvider<T, TAlarmConfig> attributeProvider)
-        {
-            _tableSource = tableSource;
-            _dimensions = dimensionProvider;
-            _attributes = attributeProvider;
-        }
-
-        private async Task<Threshold> ExpandThreshold(T resource, TAlarmConfig config, Threshold threshold)
-        {
-            if (threshold.ThresholdType == ThresholdType.PercentageOf)
-            {
-                var fraction = threshold.Value / 100;
-                var property = await _attributes.GetValue(resource, config, threshold.SourceAttribute);
-
-                threshold = new Threshold
-                {
-                    ThresholdType = ThresholdType.Absolute,
-                    Value = fraction * (double) property
-                };
-            }
-
-            return threshold;
-        }
-
-        private IList<AlarmDefinition> ExpandDefaultAlarmsForResource(IList<AlarmDefinition> alerts,
-            params Dictionary<string, AlarmValues>[] thresholds)
-        {
-            return alerts
-                .Select(x => AlarmWithMergedValues(x, thresholds))
-                .ToList();
-        }
-
-        private static AlarmDefinition AlarmWithMergedValues(
-            AlarmDefinition alarm,
-            Dictionary<string, AlarmValues>[] thresholds)
-        {
-            var mergedThreshold = MergeValueOverrides(alarm.Name, thresholds);
-
-            var copy = alarm.Copy();
-
-            copy.Threshold = new Threshold
-            {
-                SourceAttribute = alarm.Threshold.SourceAttribute,
-                ThresholdType = alarm.Threshold.ThresholdType,
-                Value = mergedThreshold.Threshold ?? alarm.Threshold.Value
-            };
-
-            copy.EvaluationPeriods = mergedThreshold.EvaluationPeriods ?? alarm.EvaluationPeriods;
-
-            copy.ExtendedStatistic = !string.IsNullOrEmpty(mergedThreshold.ExtendedStatistic)
-                ? mergedThreshold.ExtendedStatistic
-                : alarm.ExtendedStatistic;
-
-            return copy;
-        }
-
-        private static AlarmValues MergeValueOverrides(string key, Dictionary<string, AlarmValues>[] thresholds)
-        {
-            var matchesForKey = thresholds
-                .Where(t => t != null && t.ContainsKey(key))
-                .Select(t => t[key])
-                .ToList();
-
-            var matchedEvalPeriods = matchesForKey
-                .Select(t => t.EvaluationPeriods)
-                .FirstOrDefault(t => t.HasValue);
-
-            var matchedThresholdValue = matchesForKey
-                .Select(t => t.Threshold)
-                .FirstOrDefault(t => t.HasValue);
-
-            var matchedExtendedStatistic = matchesForKey
-                .Select(t => t.ExtendedStatistic)
-                .FirstOrDefault(t => !string.IsNullOrEmpty(t));
-
-            return new AlarmValues(matchedThresholdValue, matchedEvalPeriods, matchedExtendedStatistic);
-        }
-
-        public async Task<IList<Alarm>>  GenerateAlarmsFor(
-            AwsServiceAlarms<TAlarmConfig> service,
-            IList<AlarmDefinition> defaults,
-            string alarmSuffix)
-        {
-            if (service?.Resources == null || service.Resources.Count == 0)
-            {
-                return new List<Alarm>();
-            }
-
-            var allAlarms = await Task.WhenAll(service.Resources
-                .Select(r => ExpandAlarmsToResources(
-                    defaults,
-                    r,
-                    service,
-                    alarmSuffix)));
-
-            return allAlarms.SelectMany(x => x).ToList();
-        }
-
-        private async Task<IList<Alarm>> ExpandAlarmsToResources(
-            IList<AlarmDefinition> defaults,
-            ResourceThresholds<TAlarmConfig> resource,
-            AwsServiceAlarms<TAlarmConfig> service,
-            string groupSuffix)
-        {
-            // apply thresholds from resource or alerting group
-            var expanded = ExpandDefaultAlarmsForResource(defaults, resource.Values, service.Values);
-
-            var config = MergeConfiguration(service.Options, resource.Options);
-            
-            return await GetAlarms(expanded, resource, config, groupSuffix);
-        }
-
-        private string GetAlarmName(AwsResource<T> resource, string alertName, string groupSuffix)
-        {
-            return $"{resource.Name}-{alertName}-{groupSuffix}";
-        }
-
-        private async Task<IList<Alarm>> GetAlarms(IList<AlarmDefinition> alarms,
-            ResourceThresholds<TAlarmConfig> awsResource,
-            TAlarmConfig configuration,
-            string groupSuffix)
-        {
-            var result = new List<Alarm>();
-
-            var entity = await _tableSource.GetResourceAsync(awsResource.Name);
-
-            if (entity == null)
-            {
-                throw new Exception($"Entity {awsResource.Name} not found");
-            }
-
-            // expand dynamic thresholds
-            foreach (var alarm in alarms)
-            {
-                alarm.Threshold = await ExpandThreshold(entity.Resource, configuration, alarm.Threshold);
-                
-                var dimensions = _dimensions.GetDimensions(entity.Resource, alarm.DimensionNames);
-
-                var model = new Alarm
-                {
-                    AlarmName = GetAlarmName(entity, alarm.Name, groupSuffix),
-                    Resource = entity,
-                    Dimensions = dimensions,
-                    AlarmDefinition = alarm
-                };
-                result.Add(model);
-            }
-
-            return result;
-        }
-
-        private TAlarmConfig MergeConfiguration(TAlarmConfig serviceLevel, TAlarmConfig resourceLevel)
-        {
-            if (resourceLevel == null)
-            {
-                return serviceLevel ?? new TAlarmConfig();
-            }
-
-            if (serviceLevel == null)
-            {
-                return resourceLevel;
-            }
-
-            return resourceLevel.Merge(serviceLevel);
-        }
-    }
-}
